+++ src/play_motion2/play_motion2.cpp
@@ -177,8 +177,9 @@ rclcpp_action::GoalResponse PlayMotion2::handle_goal(
 }
 
 rclcpp_action::CancelResponse PlayMotion2::handle_cancel(
-  const std::shared_ptr<GoalHandlePM2>/*goal_handle*/) const
+  const std::shared_ptr<GoalHandlePM2> goal_handle) const
 {
+  RCLCPP_INFO_STREAM(get_logger(), "Cancelling motion " << goal_handle->get_goal()->motion_name);
   return rclcpp_action::CancelResponse::ACCEPT;
 }
 
@@ -189,7 +190,6 @@ void PlayMotion2::handle_accepted(const std::shared_ptr<GoalHandlePM2> goal_hand
 
 void PlayMotion2::execute_motion(const std::shared_ptr<GoalHandlePM2> goal_handle)
 {
-  auto result = std::make_shared<PlayMotion2Action::Result>();
   auto goal = goal_handle->get_goal();
 
   const double approach_time = calculate_approach_time(goal->motion_name);
@@ -204,31 +204,35 @@ void PlayMotion2::execute_motion(const std::shared_ptr<GoalHandlePM2> goal_handl
   for (const auto & [controller, trajectory] : ctrl_trajectories) {
     auto jtc_future_gh = send_trajectory(controller, trajectory);
     if (!jtc_future_gh.valid()) {
-      result->success = false;
       RCLCPP_INFO_STREAM(get_logger(), "Cannot perform motion '" << goal->motion_name << "'");
-
       // cancel all sent goals
       for (const auto & client : action_clients_) {
         client.second->async_cancel_all_goals();
       }
 
+      auto result = std::make_shared<PlayMotion2Action::Result>();
+      result->success = false;
+      result->error = "Motion " + goal->motion_name + " aborted. Cannot send goal to " + controller;
       goal_handle->abort(result);
       is_busy_ = false;
       return;
     }
     futures_list.push_back(std::move(jtc_future_gh));
   }
+
   /// @todo send feedback
-  result->success = wait_for_results(
-    futures_list,
+  auto result = wait_for_results(
+    goal_handle, futures_list,
     motions_[goal->motion_name].times.back() + extra_time);
 
-  if (!result->success) {
-    RCLCPP_INFO_STREAM(get_logger(), "Motion '" << goal->motion_name << "' failed");
-    goal_handle->abort(result);
-  } else {
-    RCLCPP_INFO_STREAM(get_logger(), "Motion '" << goal->motion_name << "' completed");
-    goal_handle->succeed(result);
+  if (goal_handle->is_active()) {
+    if (!result->success) {
+      RCLCPP_INFO_STREAM(get_logger(), "Motion '" << goal->motion_name << "' failed");
+      goal_handle->abort(result);
+    } else {
+      RCLCPP_INFO_STREAM(get_logger(), "Motion '" << goal->motion_name << "' completed");
+      goal_handle->succeed(result);
+    }
   }
   is_busy_ = false;
 }
@@ -452,7 +456,6 @@ FollowJTGoalHandleFutureResult PlayMotion2::send_trajectory(
     return {};
   }
 
-  //
   FollowJTGoalHandleFutureResult result;
   try {
     result = action_client->async_get_result(goal_handle.get());
@@ -463,11 +466,20 @@ FollowJTGoalHandleFutureResult PlayMotion2::send_trajectory(
   return result;
 }
 
-bool PlayMotion2::wait_for_results(
+PlayMotion2Action::Result::SharedPtr PlayMotion2::wait_for_results(
+  const std::shared_ptr<GoalHandlePM2> goal_handle,
   std::list<FollowJTGoalHandleFutureResult> & futures_list,
   const double motion_time)
 {
   bool failed = false;
+  auto result = std::make_shared<PlayMotion2Action::Result>();
+
+  if (goal_handle->is_canceling()) {
+    result->success = false;
+    result->error = "Motion canceled";
+    goal_handle->canceled(result);
+    return result;
+  }
 
   // Spin all futures and remove them when succeeded.
   // If one fails, set failed to true and returns false
@@ -480,14 +492,15 @@ bool PlayMotion2::wait_for_results(
           return true;
         } else {
           failed = true;
-          RCLCPP_ERROR(get_logger(), "Joint Trajectory failed");
+          result->error = "Joint Trajectory failed";
+          RCLCPP_ERROR_STREAM(get_logger(), result->error);
         }
       }
       return false;
     };
 
   // finish if failed, motions finished or timeout
-  const double TIMEOUT = motion_time * 2.0;
+  const double TIMEOUT = motion_time * 2.0 + 1.0;
   const rclcpp::Time init_time = now();
   bool on_time = true;
   do {
@@ -500,16 +513,45 @@ bool PlayMotion2::wait_for_results(
       get_controller_states(), "active", "joint_trajectory_controller/JointTrajectoryController");
 
     if (current_states != motion_controller_states_) {
+      std::string controller_name = "";
+
+      for (const auto & motion_controller : motion_controller_states_) {
+        const auto controller = std::find_if(
+          current_states.cbegin(), current_states.cend(),
+          [&](const auto & current_controller_state) {
+            return current_controller_state.name == motion_controller.name;
+          });
+
+        if (controller == current_states.end()) {
+          controller_name = motion_controller.name;
+        }
+      }
+
       failed = true;
-      RCLCPP_ERROR(get_logger(), "Controller States have changed");
+      result->error = "State of controller '" + controller_name +
+        "' has changed while executing the motion";
+      RCLCPP_ERROR_STREAM(get_logger(), result->error);
+    }
+
+    if (goal_handle->is_canceling()) {
+      // cancel all sent goals
+      for (const auto & client : action_clients_) {
+        client.second->async_cancel_all_goals();
+      }
+      futures_list.clear();
+      result->success = false;
+      result->error = "Motion canceled";
+      goal_handle->canceled(result);
     }
   } while (!failed && !futures_list.empty() && on_time);
 
-  RCLCPP_ERROR_EXPRESSION(
-    get_logger(),
-    !on_time, "Timeout exceeded while waiting for results");
+  if (!on_time) {
+    result->error = "Timeout exceeded while waiting for results";
+    RCLCPP_ERROR_STREAM(get_logger(), result->error);
+  }
 
-  return on_time && !failed;
+  result->success = on_time && !failed;
+  return result;
 }
 
 double play_motion2::PlayMotion2::calculate_approach_time(const std::string motion_key)